extract all types related to filtering to their own file for globabl use
